requirements.txt
colorama~=0.4.4
termcolor~=1.1.0
pynput~=1.7.4
pillow>=8.4.0
D3DShot>=0.1.3 -c constraints.txt